Fix list and summary for missing .bbl file.
When printing the list of files or the final summary, count in the
"virtual" .bbl file as well. This ensures the user is not surprised when
actually listing the files from the tar archive.

When printing the list, index the .bbl file with a '*' to indicate that
is generated and not taken from source dir.