Skip explicitly excluded .bbl when it is missing.
When a glob matching .bbl file name is marked as excluded using
'-x/--excl', but is missing from source dir anyway, the file is not
added to the initially globbed list of files. Instead, we explcitly
check if any of the globs would fnmatch() against <main_file.bbl>
and add <main_file.bbl> to the list if so.

Check if .bbl file is excluded before creating the BytesIO object for
it, for inclusion into tarball.