Fix tar dest when new tarfile name is entered.
When a tar file with the same name exists, the user may choose to
specify a new name for the new tar file to resolve the conflict.
However, this name must be interpreted relative to the current working
dir and not according to the main LaTeX file dir. Fixed by this commit.