Determine compression from target name if possible
Allow user to specify extensions as part of target tar file names, and
use extension to determine tar compression if it matches one of the
allowed tar.EXT names.

Ext determined from '--output' string is overridden if one of the tar
compression options is passed, with the latter taking precedence.

When resolving a tar file naming conflict, if the user specifies a new
name for the target tarball with an extension matching any of the tar
extensions, then the compression is again set according to this new
tarball name.