Fix: Don't show full path of .bbl file for '-l'.
When the .bbl file has to be generated in a tmp dir, i.e. it is missing
from source dir, passing the '-l' option printed a list with the
"virtual" .bbl file shown as a fake full path pointind to the source
dir. It should just show the basename. Fixed with this commit.