Fix: Set u/g for generated files from main file.
Supplementary files (.bbl/.ind) missing from srcdir but marked for
inclusion are copied as BytesIO objects into the tarball. This has the
consequence that the user and group attributes of these files in the
tarball were being set to 0/0.

Fixed by this commit by copying over the appropriate attribute from the
main tex file.

Add checks for bib/index test to assert the correct attributes are set.