Fix suffix swallowing for non-tar output path.
If the user enters "-o Foo.bar.baz", they expect the tarball name to be
"Foo.bar.baz.tar.?z", not "Foo.bar.tar.?z". We were getting the latter
due to the use of "Path().with_suffix('.tar')" which removes the last
suffix in Path irrespective of what it is. Fixed by simply appending
.tar to the name instead of suffix swallowing.