feat: Show spinner when compiling LaTeX project.
LaTeX compilation, especially with lualatex, can take a while depending
on the complexity of the tex files. Instead of showing a no-feedback
blank space on the terminal whilst compilation proceeds, show a spinner
with a message, so the user has clear feedback on what is going on.

Adds dependency on 'rich'.