test: Reorganise fixture for better reusability.
For the default_tartex_obj fixture, pass a lambda that accepts a file
name to use for the main tex file instead of passing the fully formed
TarTeX object. This allows us to drop defining an additional, quite
similar looking, fixture for the noext test, wherein we can just set a
filename without the extension for the lambda function.