feat: Option `--with-pdf` to add output pdf to tar
This will either directly add the final pdf output to the tarball if it
already exists in source dir or, if compilation is forced/required, from
the compiled dir.

If LaTeX compilation is not forced or required, but the final pdf cannot
be found in source dir either, it will not be added to the final tarball
and a warning will be logged.