fix: add context mgr to handle git working tree.
Add a new contextmanager `git_checkout` that handles checking out of
revisions in a git working tree and restoring it back. This is a working
attempt at leaving the user's git tree un-altered, but we could try to
improve this later by directly using file spec SHA's and `git cat-file`.