fix: don`t restore tree if git checkout fails.
Catch exception when initial `git checkout` fails, log critical,
re-raise error. This is typical of the case when the working tree has
uncommitted modifications, that is, it is unclean.

Move final restoration of git tree out of `finally` block, to prevent
attempts to restore git tree when exception is caught, as this was
over-writing uncommitted changes (uses `-f`).