chore: side-step add and excl files when using git
Avoid logging list of files to add/excl and generally simply do not
enter main function bodies for addition and exclusions of file patterns.
When using `--git-rev`, use all and only those files which are tracked
in the git tree. User must ensure that nothing important is left
untracked in the working tree.