fix: check files against ref tarball objects.
This considerably improves the robustness of `--check`, esp. when used
together with `--git-rev` by setting up a reference tarball, populating
it, and comparing its objects to a dummy target tarball which is
populated by objects from an explicit recompilation.

When `--git-rev` is not used, the check simply ensures that the
reference tarball objects are all available in source tree. No second
tarball is set up.