refactor: check target tar against minimal tar.
This is a re-working of the `--check` logic.

A `--check` should inform the user whether their intended target tarball
will contain at least the files needed to re-compile their LaTeX project
and whether it will end up containing more than this minimal content. It
should inform users whether any files they manually choose to exclude
will end up omitting objects from the tarball necessary for a
recompilation. Likewise, on the other hand, it should inform them if
their choice of `-a` or `-b` options will add files to the tarball that
may be superfluous to compilation requirements.

To this end, we set up the check as follows:

* Generate a minimal Tarballer object by explicitly re-compiling the
  project and noting the inputs required. If `latexmk` fails at this
  stage, this is likely because of files mandatory for recompilation
  missing from the source tree — deleted by accident, perhaps — and the
  check is automatically deemed to have **failed**. Note that this
  "reference" tarballer object does not include, nor exclude any user
  specified files.
* Generate a target tarball respecting user options. This will include
  any additional files marked by the user to added or excluded, such as
  bib files (`-b`).
* Compare reference and target tarball. If there are files in the
  former that are missing from the latter, the target tarball contents
  will not compile successfully in a standalone dir and the check is
  deemed to have **failed**. Report missing files. If the latter has
  files the former does not, warn the user about extra files superfluous
  to recompilation (but don't consider test to have failed).
* If the test does not fail, mark check to have passed. Report on
  successfully found minimal list of files and additional, unnecessary
  to compilation files found in target and return.