feat: modify `--check` so it generates output tar.
Modify the `--check` option to allow users to generate tarballs after
running a check on whether the minimal set of files required for
compilation will be included. The check in this case is run in "silent"
mode, only reporting on errors due to missing/excluded essential files,
and raising a `SystemExit` if it finds any. Warnings due to unnecessary
files being added to tarball and check success messages are all
suppressed.

Add an `--only-check` option to do what `--check` was doing before,
viz., running a verbose check and reporting on missing, unnecessary, and
success status of the check without actually generating the tarball.