fix: save missing supplementary filenames to cache
When needed but missing supplementary files (.bbl, .idx) are added as
streams to the tarball using a recompile, store their names in the cache
file with the new "streams" key. The presence of a file here means that
supplementary file is missing from the user's project tree and a fresh
recompile will be needed every time until it has been regenerated or
restored there (the user manually running latexmk in the project dir
should do the trick). Automatically return `False` from the hash check
if the "streams" dict has any item matching a supplementary file
extension. This will force a recompile (as it should).