fix: use param to check if cache must be updated.
In `.input_files_from_recompile()`, the cache should not be always
updated or touched. For instance, when a user forces a recompile,
using `-F` on the cmdline, this must not look at the cache data at
all. However, when a recompile is necessitated due to cache missing or
mismatching of sha sums, then it must be updated when calling
`.input_files_from_recompile()`. Add a bool parameter to the function to
handle this conditional.