fix: log warning for all input "streams".
When getting list of INPUT files from the cache, anything in "streams"
is there because the file itself is missing from the project dir.
Otherwise, it would have been cached under the "input_files" key.
Directly warn about these being missing even when the cache otherwise
tallies up with sha256sums of all other files.

Note that, there is no way to check the integrity of a "restored"
missing file, so until the cache is regenerated due to some source file
modifications, the "missing" status of the stream must continue.