feat: save cache files in XDG_CACHE_HOME subdir.
Save cache files to `XDG_CACHE_HOME/tartex` and store them in a
project dir specific unique dir generated by appending a short hash to
the project parent dir. This enables the cache dir to be user
identifiable, because of the use of the parent dir lead, as well as
unique due to the use of a short hash appended to it.