test: don't use capsys msg to check num of files.
Directly assert against the length of `tarfile.getnames()` from the
generated tarball instead.

Capsys/caplog output, handled by `rich`, may carry formatting marks
based on whether output is captured to a supported console or not.
Asserts based on msg content may or may not work depending on the
conditions under which `pytest` is run. Avoid it.