fix: don't send empty lists to `TarBaller`.
For cmdline options `-b`/`-a`, it may be the case that the list of files
to be actually added may turn out to be empty. In the former case, this
would be because the project may not actually be using bibtex, for
instance. For option `-a`, files matching the user supplied globs may
not actually exist in the source dir. In all these cases, the generated
list will turn out to be empty, and must not be sent to
`TarBaller.add_files()`.