refactor: use helper func to filter TeX err/warns.
Define a helper function that accepts patterns to match against and
replace with as a parameter and filters them out from the full latexmk
error output. An optional list of strings may be passed as the `ignore`
param to ignore an error or warning if matched patterns is in this list.