feat: different exit codes based on error origin.
Add an `IntEnum` class to represent different exit codes based on the
source of the exception. This will allow us to "signal" to applications
the reason for tartex's termination, allowing for clearer end-user
messages and the like.

For now, only use a different exit code for latexmk errors, since these
typically save a log file, whereas none of the others have any
non-transient outputs.